refactor: update all hydration functions to be named as such
Updated the symbols for the hydration functions. Tested imports agains processDeRecord.
