+++ apps/db_management/src/models/german/hydrationFunctions/hydrateKonjunktivConjugation.ts
